Put pagination stuff in Pagination module
And put class in its own file
