+++ lib/elastic_record/index.rb
@@ -3,6 +3,8 @@ require 'elastic_record/index/deferred'
 require 'elastic_record/index/documents'
 require 'elastic_record/index/manage'
 require 'elastic_record/index/mapping'
+require 'elastic_record/index/pagination'
+require 'elastic_record/index/scroll_enumerator'
 require 'elastic_record/index/search'
 require 'elastic_record/index/settings'
 require 'elastic_record/index/mapping_type'
@@ -28,7 +30,7 @@ module ElasticRecord
   # [update_mapping]
   #   Update elastic search's mapping
   class Index
-    include Documents, Search
+    include Documents, Search, Pagination
     include Manage
     include Mapping, Settings
     include Analyze
